# docs/swagger.yaml
definitions:
  internal_controllers_webhook.GenericResponse:
    properties:
      message:
        description: Message provides a brief status message for the operation.
        type: string
    type: object
  internal_controllers_webhook.RegisterWebhookRequest:
    properties:
      condition:
        description: Condition is a CEL expression evaluated against the metric to
          decide when to fire.
        example: value > 55
        type: string
      coolDownPeriod:
        description: CoolDownPeriod is the minimum number of seconds between successive
          firings.
        example: 30
        type: integer
      description:
        description: Description is an optional human-friendly explanation of the
          webhook.
        example: This webhook is used to notify when the speed of the vehicle exceeds
          55 mph.
        type: string
      displayName:
        description: DisplayName is a user-friendly unique name per developer license.
        example: Speed Alert
        type: string
      metricName:
        description: MetricName is the fully qualified signal/metric to monitor.
        example: speed
        type: string
      service:
        description: Service is the subsystem producing the metric (e.g. "vehicles").
        example: telemetry.signals
        type: string
      status:
        description: Status sets the initial state for the webhook (e.g. "enabled"
          or "Disabled").
        example: enabled
        type: string
      targetURL:
        description: TargetURL is the HTTPS endpoint that will receive webhook callbacks.
        example: https://example.com/webhook
        type: string
      verificationToken:
        description: VerificationToken is the expected token that your endpoint must
          echo back during verification.
        example: "1234567890"
        type: string
    required:
    - condition
    - coolDownPeriod
    - metricName
    - service
    - targetURL
    - verificationToken
    type: object
  internal_controllers_webhook.RegisterWebhookResponse:
    properties:
      id:
        description: ID is the unique identifier of the created webhook.
        type: string
      message:
        description: Message provides a brief status message for the operation.
        type: string
    type: object
  internal_controllers_webhook.SignalDefinition:
    properties:
      description:
        description: Description briefly explains what the signal represents.
        type: string
      name:
        description: Name is the JSON-safe name of the signal.
        type: string
      unit:
        description: Unit is the unit of measurement for the signal value (if any).
        type: string
    type: object
...
  internal_controllers_webhook.UpdateWebhookRequest:
    properties:
      condition:
        description: Condition updates the CEL expression used to decide when to fire.
        type: string
      coolDownPeriod:
        description: CoolDownPeriod updates the minimum number of seconds between
          firings.
        type: integer
      description:
        description: Description updates the optional human-friendly explanation of
          the webhook.
        type: string
      displayName:
        description: DisplayName updates the user-friendly unique name per developer
          license.
        type: string
      metricName:
        description: MetricName updates the signal/event name used by the webhook.
        type: string
      status:
        description: Status updates the current state of the webhook (e.g. "enabled"
          or "Disabled").
        type: string
      targetURL:
        description: TargetURL updates the HTTPS endpoint that will receive callbacks.
        type: string
    type: object
  internal_controllers_webhook.UpdateWebhookResponse:
    properties:
      id:
        description: ID is the unique identifier of the updated webhook.
        type: string
      message:
        description: Message provides a brief status message for the operation.
        type: string
    type: object
  internal_controllers_webhook.WebhookView:
    properties:
      condition:
        description: Condition is the CEL expression evaluated to decide when to fire.
        type: string
      coolDownPeriod:
        description: CoolDownPeriod is the minimum number of seconds between successive
          firings.
        type: integer
      createdAt:
        description: CreatedAt is when the webhook was created.
        type: string
      description:
        description: Description is an optional human-friendly explanation of the
          webhook.
        type: string
      displayName:
        description: DisplayName is the user-friendly unique name per developer license.
        type: string
      failureCount:
        description: FailureCount counts consecutive delivery failures for observability.
        type: integer
      id:
        description: ID is the unique identifier of the webhook.
        type: string
      metricName:
        description: MetricName is the fully qualified signal/metric monitored by
          the webhook.
        type: string
      service:
        description: Service is the subsystem producing the metric (e.g. "vehicles").
        type: string
      status:
        description: Status is the current state of the webhook (e.g. "enabled" or
          "Disabled").
        type: string
      targetURL:
        description: TargetURL is the HTTPS endpoint that receives webhook callbacks.
        type: string
      updatedAt:
        description: UpdatedAt is when the webhook was last modified.
        type: string
    type: object
info:
  contact: {}
  title: Vehicle Triggers API
  version: "1.0"
paths:
  /v1/webhooks:
    get:
      description: Retrieves all registered webhooks for the developer.
      produces:
      - application/json
      responses:
        "200":
          description: List of webhooks
          schema:
            items:
              $ref: '#/definitions/internal_controllers_webhook.WebhookView'
            type: array
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: List all webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Registers a new webhook with the specified configuration. The target
        URI is validated to ensure it is a valid URL, responds with 200 within a timeout,
        and returns a verification token.
      parameters:
      - description: Webhook configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers_webhook.RegisterWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Webhook registered successfully
          schema:
            $ref: '#/definitions/internal_controllers_webhook.RegisterWebhookResponse'
        "400":
          description: Invalid request payload or target URI
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Register a new webhook
      tags:
      - Webhooks
  /v1/webhooks/{webhookId}:
    delete:
      description: Deletes a webhook by its ID.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Webhook deleted successfully
          schema:
            $ref: '#/definitions/internal_controllers_webhook.GenericResponse'
        "404":
          description: Webhook not found
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Delete a webhook
      tags:
      - Webhooks
    get:
      description: Returns every vehicleTokenId currently subscribed.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all vehicles subscribed to a webhook
      tags:
      - Webhooks
    put:
      consumes:
      - application/json
      description: Updates the configuration of a webhook by its ID. The failure count
        is reset to 0 when updating a webhook.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      - description: Webhook configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers_webhook.UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook updated successfully
          schema:
            $ref: '#/definitions/internal_controllers_webhook.UpdateWebhookResponse'
        "400":
          description: Invalid request payload
        "404":
          description: Webhook not found
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Update a webhook
      tags:
      - Webhooks
...
  /v1/webhooks/signals:
    get:
      description: Fetches the list of signal names available for the data field.
      produces:
      - application/json
      responses:
        "200":
          description: List of signal names
          schema:
            items:
              $ref: '#/definitions/internal_controllers_webhook.SignalDefinition'
            type: array
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get signal names
      tags:
      - Webhooks
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"